Clean up build warnings

+++ Pod/Classes/ENHViewController.m
@@ -538,9 +538,9 @@ static const CGFloat __blurTintColorAlpha = 0.2f;				// defines how much to tint
             }
             
             // angle (θ) is the angle between the push vector (V) and vector component parallel to radius, so it should always be positive
-            CGFloat angle = fabsf(fabsf(velocityAngle) - fabsf(locationAngle));
+            CGFloat angle = fabs(fabs(velocityAngle) - fabs(locationAngle));
             // angular velocity formula: w = (abs(V) * sin(θ)) / abs(r)
-            CGFloat angularVelocity = fabsf((fabsf(pushVelocity) * sinf(angle)) / fabsf(radius));
+            CGFloat angularVelocity = fabs((fabs(pushVelocity) * sinf(angle)) / fabs(radius));
             
             // rotation direction is dependent upon which corner was pushed relative to the center of the view
             // when velocity.y is positive, pushes to the right of center rotate clockwise, left is counterclockwise
@@ -551,8 +551,8 @@ static const CGFloat __blurTintColorAlpha = 0.2f;				// defines how much to tint
             // amount of angular velocity should be relative to how close to the edge of the view the force originated
             // angular velocity is reduced the closer to the center the force is applied
             // for angular velocity: positive = clockwise, negative = counterclockwise
-            CGFloat xRatioFromCenter = fabsf(offsetFromCenter.horizontal) / (CGRectGetWidth(self.imageView.frame) / 2.0f);
-            CGFloat yRatioFromCetner = fabsf(offsetFromCenter.vertical) / (CGRectGetHeight(self.imageView.frame) / 2.0f);
+            CGFloat xRatioFromCenter = fabs(offsetFromCenter.horizontal) / (CGRectGetWidth(self.imageView.frame) / 2.0f);
+            CGFloat yRatioFromCetner = fabs(offsetFromCenter.vertical) / (CGRectGetHeight(self.imageView.frame) / 2.0f);
             
             // apply device scale to angular velocity
             angularVelocity *= deviceAngularScale;
